refactor(workout-plan): extract superset actions (WP3.4g)
